+++ docs/src/app/components/pages/components/Card/ExampleControlled.js
@@ -34,7 +34,7 @@ export default class CardExampleControlled extends React.Component {
         <CardHeader
           title="URL Avatar"
           subtitle="Subtitle"
-          avatar="http://lorempixel.com/100/100/nature/"
+          avatar="images/ok-128.jpg"
           actAsExpander={true}
           showExpandableButton={true}
         />
@@ -50,7 +50,7 @@ export default class CardExampleControlled extends React.Component {
           expandable={true}
           overlay={<CardTitle title="Overlay title" subtitle="Overlay subtitle" />}
         >
-          <img src="http://lorempixel.com/600/337/nature/" />
+          <img src="images/nature-600-337.jpg" />
         </CardMedia>
         <CardTitle title="Card title" subtitle="Card subtitle" expandable={true} />
         <CardText expandable={true}>